PollingQuery: Rename java files and implement publisher
