+++ examples/polling_querycondition/.gitignore
@@ -1,3 +1,5 @@
 # Files generated by rtiddsgen
 flights.*
 !flights.idl
+Flight.*
+!Flight.idl
