+++ examples/polling_querycondition/java/USER_QOS_PROFILES.xml
@@ -9,18 +9,18 @@
  any incidental or consequential damages arising out of the use or inability to
  use the software.
  -->
-<!-- 
+<!--
 Description
 XML QoS Profile for waitset_statuscond
 
-The QoS configuration of the DDS entities in the generated example is 
+The QoS configuration of the DDS entities in the generated example is
 loaded from this file.
 
 This file is used only when it is in the current working directory
 or when the enviroment variable
 NDDS_QOS_PROFILES is defined and points to this file.
 
-For more information about XML QoS Profiles see Chapter 15 in the 
+For more information about XML QoS Profiles see Chapter 15 in the
 RTI Connext user manual.
 -->
 <dds xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
@@ -32,46 +32,24 @@ RTI Connext user manual.
     -->
     <qos_library name="waitset_statuscond_Library">
 
-        <!-- QoS profile used to configure reliable communication between the DataWriter 
+        <!-- QoS profile used to configure reliable communication between the DataWriter
              and DataReader created in the example code.
 
              A QoS profile groups a set of related QoS.
         -->
         <qos_profile name="waitset_statuscond_Profile" is_default_qos="true">
-            <!-- QoS used to configure the data writer created in the example code -->                
+            <!-- QoS used to configure the data writer created in the example code -->
             <datawriter_qos>
                 <reliability>
                     <kind>RELIABLE_RELIABILITY_QOS</kind>
-                    <max_blocking_time>
-                        <sec>60</sec>
-                    </max_blocking_time>
-                </reliability>                
-
-                <history>
-                    <kind>KEEP_ALL_HISTORY_QOS</kind>
-                </history>
-
-                <protocol>
-                    <rtps_reliable_writer>
-                        <min_send_window_size>50</min_send_window_size>
-                        <max_send_window_size>50</max_send_window_size>
-                    </rtps_reliable_writer>
-                </protocol>
-		
+                </reliability>
             </datawriter_qos>
 
-            <!-- QoS used to configure the data reader created in the example code -->                
+            <!-- QoS used to configure the data reader created in the example code -->
             <datareader_qos>
-
                 <reliability>
                     <kind>RELIABLE_RELIABILITY_QOS</kind>
                 </reliability>
-    
-                <history>
-                    <kind>KEEP_LAST_HISTORY_QOS</kind>
-                    <depth>6</depth>
-                </history>
-			
             </datareader_qos>
 
             <participant_qos>
